+++ packages/ui/src/Avatar/Avatar.tsx
@@ -1,0 +1,9 @@
+import { forwardRef } from 'react';
+import MuiAvatar, { type AvatarProps } from '@mui/material/Avatar';
+export { default as AvatarGroup, type AvatarGroupProps } from '@mui/material/AvatarGroup';
+
+export type { AvatarProps };
+
+export const Avatar = forwardRef<HTMLDivElement, AvatarProps>(function Avatar(props, ref) {
+  return <MuiAvatar ref={ref} {...props} />;
+});